requirements.txt
certifi==2025.8.3
charset-normalizer==3.4.3
flake8==7.3.0
idna==3.10
mccabe==0.7.0
proxmoxer==2.2.0
pycodestyle==2.14.0
pyflakes==3.4.0
python-dotenv==1.1.1
requests==2.32.5
urllib3==2.5.0
